add remaining character user relations

+++ RpgWebApi/Services/CharacterService/CharacterService.cs
@@ -7,33 +7,49 @@ using System.Threading.Tasks;
 using RpgWebApi.Data;
 using RpgWebApi.Dtos.Character;
 using RpgWebApi.Models;
+using Microsoft.AspNetCore.Http;
+using System.Security.Claims;
 
 namespace RpgWebApi.Services.CharacterService
 {
     public class CharacterService : ICharacterService
     {
-        //private static List<Character> characters = new List<Character>{
-        //    new Character(),
-        //    new Character(){ Id = 1, Name = "Sam"}
-        //};
-
         private IMapper _mapper;
         private DataContext _context;
+        private IHttpContextAccessor _accessor;
 
-        public CharacterService(IMapper mapper, DataContext context)
+        public CharacterService(IMapper mapper, DataContext context, IHttpContextAccessor accessor)
         {
             _mapper = mapper;
             _context = context;
+            _accessor = accessor;
         }
+
+        private int GetUserId() => int.Parse(_accessor.HttpContext.User.FindFirstValue(ClaimTypes.NameIdentifier));
 
         public async Task<ServiceResponse<List<GetCharacterDto>>> AddCharacter(AddCharacterDto newCharacter)
         {
             var serviceResponse = new ServiceResponse<List<GetCharacterDto>>();
-            Character character = _mapper.Map<Character>(newCharacter);
 
-            await _context.Characters.AddAsync(character);
-            await _context.SaveChangesAsync();
-            serviceResponse.Data = _mapper.Map<List<GetCharacterDto>>(await _context.Characters.ToListAsync());
+            try
+            {
+                Character character = _mapper.Map<Character>(newCharacter);
+
+                character.User = await _context.Users.FirstOrDefaultAsync(u => u.Id == GetUserId());
+
+                await _context.Characters.AddAsync(character);
+                await _context.SaveChangesAsync();
+
+                serviceResponse.Data = _mapper.Map<List<GetCharacterDto>>(
+                        await _context.Characters.Where(c => c.User.Id == GetUserId()).ToListAsync()
+                        );
+            }
+            catch (Exception e)
+            {
+                serviceResponse.Message = e.Message;
+                serviceResponse.Success = false;
+            }
+
             return serviceResponse;
         }
 
@@ -41,30 +57,49 @@ namespace RpgWebApi.Services.CharacterService
         {
             var serviceResponse = new ServiceResponse<List<GetCharacterDto>>();
 
-            try
+            try 
             {
-                var deletedCharacter = await _context.Characters.FirstOrDefaultAsync(c => c.Id == id);
+                var deletedCharacter = await _context.Characters.FirstOrDefaultAsync(c => c.Id == id && c.User.Id == GetUserId());
 
-                _context.Characters.Remove(deletedCharacter);
-                await _context.SaveChangesAsync();
+                if (deletedCharacter != null)
+                {
+                    _context.Characters.Remove(deletedCharacter);
+                    await _context.SaveChangesAsync();
 
-                serviceResponse.Data = _mapper.Map<List<GetCharacterDto>>(await _context.Characters.ToListAsync());
+                    serviceResponse.Data = _mapper.Map<List<GetCharacterDto>>(
+                        await _context.Characters.Where(c => c.User.Id == GetUserId()).ToListAsync()
+                        );
+                }
+                else
+                { 
+                    serviceResponse.Message = "no such character";
+                    serviceResponse.Success = false;
+                }
             }
             catch (Exception e)
             {
                 serviceResponse.Message = e.Message;
                 serviceResponse.Success = false;
             }
+
             return serviceResponse;
         }
 
-        public async Task<ServiceResponse<List<GetCharacterDto>>> GetAllCharacters(int userId)
+        public async Task<ServiceResponse<List<GetCharacterDto>>> GetAllCharacters()
         {
             var serviceResponse = new ServiceResponse<List<GetCharacterDto>>();
 
-            var dbCharacters = await _context.Characters.Where(c => c.User.Id == userId).ToListAsync();
+            try
+            {
+                var dbCharacters = await _context.Characters.Where(c => c.User.Id == GetUserId()).ToListAsync();
 
-            serviceResponse.Data = _mapper.Map<List<GetCharacterDto>>(dbCharacters);
+                serviceResponse.Data = _mapper.Map<List<GetCharacterDto>>(dbCharacters);
+            }
+            catch (Exception e)
+            {
+                serviceResponse.Message = e.Message;
+                serviceResponse.Success = false;
+            }
 
             return serviceResponse;
         }
@@ -73,9 +108,25 @@ namespace RpgWebApi.Services.CharacterService
         {
             var serviceResponse = new ServiceResponse<GetCharacterDto>();
 
-            var dbCharacter = await _context.Characters.FirstOrDefaultAsync(c => c.Id == id);
+            try
+            { 
+                var dbCharacter = await _context.Characters.FirstOrDefaultAsync(c => c.Id == id && c.User.Id == GetUserId());
 
-            serviceResponse.Data = _mapper.Map<GetCharacterDto>(dbCharacter);
+                if (dbCharacter != null)
+                {
+                    serviceResponse.Data = _mapper.Map<GetCharacterDto>(dbCharacter);
+                }
+                else
+                {
+                    serviceResponse.Message = "no such character";
+                    serviceResponse.Success = false;
+                }
+            }
+            catch (Exception e)
+            {
+                serviceResponse.Message = e.Message;
+                serviceResponse.Success = false;
+            }
 
             return serviceResponse;
         }
@@ -83,17 +134,26 @@ namespace RpgWebApi.Services.CharacterService
         public async Task<ServiceResponse<GetCharacterDto>> UpdateCharacter(UpdateCharacterDto character)
         {
             var serviceResponse = new ServiceResponse<GetCharacterDto>();
+
             try
-            {
-                var newCharacter = await _context.Characters.FirstOrDefaultAsync(c => c.Id == character.Id);
+            { 
+                var newCharacter = await _context.Characters.FirstOrDefaultAsync(c => c.Id == character.Id && c.User.Id == GetUserId());
 
-                foreach (var prop in newCharacter.GetType().GetProperties())
-                    prop.SetValue(newCharacter, character.GetType().GetProperty(prop.Name)?.GetValue(character));
+                if (newCharacter != null)
+                {
+                    foreach (var prop in newCharacter.GetType().GetProperties())
+                        prop.SetValue(newCharacter, character.GetType().GetProperty(prop.Name)?.GetValue(character));
 
-                await _context.SaveChangesAsync();
+                    await _context.SaveChangesAsync();
 
-                var returnValue = _mapper.Map<GetCharacterDto>(newCharacter);
-                serviceResponse.Data = returnValue;
+                    var returnValue = _mapper.Map<GetCharacterDto>(newCharacter);
+                    serviceResponse.Data = returnValue;
+                }
+                else
+                {
+                    serviceResponse.Message = "no such character";
+                    serviceResponse.Success = false;
+                }
             }
             catch (Exception e)
             {
